migrate_bigint変更

+++ db/migrate/20230604133817_create_equipment.rb
@@ -2,7 +2,7 @@ class CreateEquipment < ActiveRecord::Migration[6.1]
   def change
     create_table :equipment do |t|
       
-      t.integer :register_id,       null: false
+      t.bigint :register_id,       null: false
       t.integer  :reference_number, null: false
       t.string  :introduction,      null: false
       t.string  :location,          null: false
